Finished Problem 60
Original algorithm found the solution in approximately 3 minutes and 30 seconds.
After a small adjustment to the PrimeTester class, time was improved to around 30 seconds.
After around round of optimizations, the runtime was down to about 3 seconds. The last optimizations were to check combinations early to see if they were valid.